Add files via upload
i cannot upload the full unity project onto this repository because it has a size of 100MB. The final project from my Artificial Intelligence class at SPCS was a tank game. There are 4 other tanks that are unique and they work intelligently, together, and alone sometimes with the intention of "killing" you. This project uses Unity, Visual Studio, and Panda BT.